Preservation: Stabilize SunOS Bourne sh on 64-bit Linux
This commit addresses critical architectural bugs discovered during the
porting of the SunOS Bourne shell's original sbrk-based memory allocator
to modern 64-bit Linux/x86_64.

Resolutions:
- Pointer Truncation: Standardized the sh_sbrk/setbrk interface to use
  'long' for increments, preventing 32-bit wrap-around and infinite
  allocation loops in blok.c and stak.c.
- ABI Consistency: Fully synchronized prototypes in 'defs.h' to prevent
  implicit 'int' return truncation of 64-bit pointers (notably for
  sh_setenv, macro, and locstak). This resolved EFAULT during execve.
- Alignment: Aligned internal stack structures by updating BYTESPERWORD
  to 8 in 'mode.h', ensuring stable data access on modern CPUs.
- Signal Safety: Updated fault handling to use POSIX sigaction with
  SA_RESTART for 64-bit stability.
- Environment Hardening: Synchronized countnam logic with pushnam to
  prevent environment array overflows.

Verified with recursive_stress_sunos.sh (Depth 1000).

+++ setbrk.c
@@ -18,7 +18,7 @@ static	char sccsid[] = "@(#)setbrk.c 1.1 94/10/31 SMI"; /* from S5R3.1 1.8 */
 
 
 char *
-setbrk(unsigned int incr)
+setbrk(long incr)
 {
 
 	register char *a = sh_sbrk(incr);
